removed deprecated hooks (#304)
