+++ packages/react-checkout/src/hooks/index.ts
@@ -5,7 +5,6 @@ export * from './useModalTheme'
 export * from './useCheckoutWhitelistStatus'
 export * from './useSardineClientToken'
 export * from './useSelectPaymentModal'
-export * from './useClearCachedBalances'
 export * from './useTransferFundsModal'
 export * from './useTransactionStatusModal'
 export * from './useSwapModal'
