+++ packages/react-wallet/src/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './useOpenWalletModal'
-export * from './data'
 export * from './useNavigation'
 export * from './useSettings'
